refactor: add model computer and repository for computer
