+++ Program.cs
@@ -1,7 +1,8 @@
 ﻿using Microsoft.Data.Sqlite; //importar o sqlite
 using LabManager.Database;
+using LabManager.Repositories;
 
-var databasesetup = new DatabaseSetup(); //instancia o database e já executa os método
+var databaseSetup = new DatabaseSetup(); //instancia o database e já executa os método
 
 // See https://aka.ms/new-console-template for more information
 //Console.WriteLine(args);
@@ -17,23 +18,14 @@ var ModelAction = args[1];
 
 if(modelName == "Computer")
 {
+    var computerRepository = new ComputerRepository();
     if(ModelAction == "List")
     {
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open(); //ABRIR O ARQUIVO/conexão database.db
-
-
-        var command = connection.CreateCommand(); //comando criado no banco aberto
-        command.CommandText = "SELECT * FROM Computers";
-
-        var reader = command.ExecuteReader(); //representa o resultado da tabela
-
-        while(reader.Read())
+        Console.WriteLine("Computer List");
+        foreach (var computer in computerRepository.GetAll())
         {
-            Console.WriteLine("{0}, {1}, {2}", reader.GetInt32(0), reader.GetString(1), reader.GetString(2)); //pegar o valor, readers em objeto Computador
+            Console.WriteLine($"{computer.Id}, {computer.Ram}, {computer.Processor}");            
         }
-
-        connection.Close(); // fecha a conexão
     }
 
     if(ModelAction == "New")
